👌 IMPROVE: animate with composable / hook
