+++ package.json
@@ -11,6 +11,7 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@vueuse/core": "10.4.1",
     "echarts": "5.4.3",
     "gsap": "3.12.2",
     "vue": "^3.3.4",
